spj-svs: Refactored SPJ API, SE-handler interface, and handling of resultsets created by SPJ.
Rational for this rewrite is that the 'global cursor' on the NdbQuery has proved insufficient for handling
a pushed join consisting of scan - scan - See added MTR testcase.

In this commit we introduce 'local cursors' on each NdbQueryOperation. These may be itterated to 
'first' or 'next' position. All navigation on these local cursors are with respect to any parent
dependencies, such that we will get an 'EndOfScan' status when there are no more 'next' results
on an operations which depends on the current parent(s).

In order to clean up the code, the global cursor has been reimplemented on top of the local cursors.
Furthermore, logic for elimination of NULL rows in inner joins and imcomplete batches, has been moved
from the resultset iteration into ::handleBatchComplete() which now filters away NULL rows before
any itteration is done on the resultset.
This enabled us to remove the enum states 'IterState' and 'ScanRowResults'.

SE-handler interface has been extended with ::index_next_pushed() which provides the interface 
to 'next' itterating on the local cursor

NDB handler implementation has been changed to use the local cursors instead of the global cursor.