BUG#56882: binlog_base64_flag fails on valgrind and win debug            max platforms
In some cases, the variable err in mysql_client_binlog_statement
would not be initialized but still some conditional jumps would
rely on its value. This is for instance the case when a BINLOG
event is truncated and it returns an error. In this case the
'err' variable is not initialized at that time, but we jump to
the label 'end' and then check the following condition: 
      
   if ((error || err) && ... )
      
This makes valgrind complain that there is a conditional jump or
move depending on an uninitialized value.

We fix this by initializing the variable err to 0.