This commit fixes a bug that caused results for pushed queries to be sorted wrongly in some cases. This could happen if the root operation was a sorted scan (of an ordered index) and there was at least one scan descendant. Then it might happen that the child batch for fragment 0 contained child tuples of parent tuples A and B, while fragment 1 only containing tuples that are children of A *with more such tuples coming in the next batch*. The API would then emit joined tuples containing parent tuple B before asking for the next batch of child tuples. Sorting order would then be wrong, since the next batch would generate more joined tuples containing parent tuple A.
The fix has two element:
1. The API will return an error code if an application tries to build
a query with a sorted root scan *and* another scan.
2. The ndb handler will not try to push such operation trees. This would 
otherwise happen if e.g. the query had an "ORDER BY" clause or if (parts of)
a predicate would map to a range scan on the root table. (In the latter case
sorting is not really needed and should ideally be eliminated.)