Reorg dict_stats_table_check()
Return the error message to the caller instead of printing it to
stderr. Also move the logic of not printing anything when table does not
exist to the caller.

This is done in order to make dict_stats_table_check() more generic and to
expose it to the outside world [1].

Suggested by:	Jimmy (rb://373) [1]