BUG#56912: binlog_killed fails sporadically on            mysql-next-mr-bugfixing codebase.
Two new event types were added recently. This made the FD event
to grow 2 bytes, thence the event positions in the binary log
shifted two bytes as well. Given that sometimes the assertion in
the binlog_killed test relies on absolute binary log positions,
it will try to read events from a current obsolete position (175,
while the new one is 177).

We fix this by making the test to read the current position from
SHOW MASTER STATUS into $start_pos and then issuing mysqlbinlog
with --start-position=$start_pos instead of a hardcoded absolute
position. This will also make the dump to read the BEGIN event
that precedes the actual INSERT event (whereas before it would
not), but this should not affect the assertion correctness.