spj-svs: mysqld wil try to avoid sending unnecessary lookup operations when the key to be looked up is the same as the key of the previous record it retrieved from the SE.
However, when the lookup operation is a root of a pushed join, there may be child operations inside this pushed join
which also depends on (through paramValues) field values from operations outside the pushed join. It is therefore not
correct to assume that the query result is 'constant' if the lookup key is the same.

To handle these cases, handler::test_push_flag(HA_PUSH_MULTIPLE_DEPENDENCY) has been introduced to 
detect when a pushed query has such child dependencies. join_no_more_records() will then invalidate 
and results from the pushed query, which force a new execution of the pushed join when another (root) lookup
is requested.