BUG#46110: --replicate-wild-do-table invalidates            --replicate-do-db matching rule
When using --replicate-do-db and --replicate-ignore-db with
--replicate-wild-do-table, some database level statements, eg
DROP DATABASE or CREATE DATABASE, might not be replayed even if
the DATABASE name was inline with the rules set in
--replicate-do-db and/or --replicate-ignore-db. The following
example would not replicate correctly:
      
  SLAVE, configured with: 
  --replicate-do-db=dbx --replicate-wild-do-table=db%.t1
      
  MASTER, dba issues:
  MASTER> CREATE DATABASE dbx;
      
Although, at first sight, there is nothing in the filter rules
that might prevent this statement to be replayed, the slave would
not execute it. In fact, "replicate-wild-do-table=db%.t1" 
invalidates the "OK" from "replicate-do-db=dbx", making the event
to be skipped/filtered out.
      
We fix this, by making --replicate-do/ignore-db rules take
precedence over replicate-wild-do-table for *database* level
statements only:
      
  - CREATE/DROP/ALTER DATABASE ...
      
As such, when replicating such statements, we only look into
replicate-wild-do-table if and only if replicate-do/ignore-db 
rules are void. Otherwise, --replicate-do/ignore-db rules 
dictate the fate of the statement: either filtered out or 
executed.