spj-svs: Added missing handling NULL values in pruned scan keys.
Pruned keys may be constructed from linkedValues which may contain a NULL value. As a
pruned key implies equality on the fields in the key, a NULL value in the prune key can
never match any existing tuple on the datanodes.

We can therefore eliminate any pruned SCANREQ with NULL value in the prune key.

In addition to the fix above, this commit also adds:

- Checking for errors returns from ::expand().
- add ndbrequire(!hasNull) to those places where we do not expect (and does not handle) 
  NULL values in the key produced by ::expand().
- Some more DEBUG output for debugging pruned scans.
- Moves the 'prune pattern' (SI_PRUNE_PATTERN) to be the last part of the 'optional'
  part of a Qserialized ueryTree node - Required as 'parent' relationship for 
  the node should be established by ::parseDA() before we could ::expand() 
  the pattern for the pruned key.