Bug #51639  Some replication filters are case sensitive, some are not!
There is an inconsistency in name comparison of rpl filters.
Some are case sensitive, while others are not. And they did
not follow the setting of lower_case_table_names.

To fix the problem to make all the filters follow the setting of
lower_case_table_name to be case sensitive and accent sensitive
when setting lower_case_table_name=0. Otherwise they will be
case insensitive but accent sensitive.