Bug#44771 : Unique Hash index in memory engine will give wrong query result for             NULL value.
In a memory table when there are more than one NULL values for a unique key
column using HASH index, a search for NULL on that column returns just one
row.

In memory engine, while reading a hash key which is NULL, it doesn't get copied
to info->lastkey, which inturn is used to search next records.

Fixed by adding the condition to check for NULL key, which lets copying of key
to info->lastkey.