BUG#57513: Remove unused and unmaintained event: Slave_log_event
The source code shows that Slave_log_event is[/has] not [been]
used at all[, ever]. So there are at least three reasons to
remove it:
 
  1. the fact that this code is unmaintained can raise some
     unexpected issues, like: BUG#38718;

  2. other unmaintained code that exhibited dependencies from
     Slave_log_event was already removed in BUG#54649

  3. Reviewers of BUG#38718 required that a new bug would be
     reported to remove this event.

To fix this, we remove the Slave_log_event associated code.  We
maintain the ENUM position for SLAVE_EVENT for historical reasons
and to avoid messing with Format_description_event
post_header_len definitions.