spj-svs: Enhanced implementation of handling 'sorted indexscan(t1) -> index ref(t2)'
We previously had to block pushing of scan childs if the root operation was an ordered index scan as
we are not able to provide these result ordered through the root operation.

This fix is now reverted - Instead we now allow these kind of queries to be pushed, but require
the result to be written to temp. file where it is filesorted. We believe the overhead of this
extra filesort to be far less than not pushing the query.