Remove all locks that are on the table or its records, if there are no refernces to the table but it has record locks, we release the record locks unconditionally. One use case is:
	CREATE TABLE t2 (PRIMARY KEY (a)) SELECT * FROM t1;

If after the user transaction has done the SELECT and there is a
problem in completing the CREATE TABLE operation, MySQL will drop
the table. InnoDB will create a new background transaction to do the
actual drop, the trx instance that is passed to this function. To
preserve existing behaviour we remove the locks but ideally we
shouldn't have to. There should never be record locks on a table
that is going to be dropped.

Add assertions to check that the table doesn't have any locks on it.

This fix should really have been part the change below. While the original
fix was correct, it forced some tables to the background drop queue that
would have been deleted immediately because of a bug in the prepatched code.
In this patch we fix that bug (partially) and preserve old behaviour.

  3262 Sunny Bains	2010-10-20
  For tables that have a reference count of zero, are not referenced
  by any foreign key and don't refer to any table (FK) are removed
  from the dictionary cache if it reaches an upper bound.