WL#5265 Deprecate features in Dahlia
This is a patch for deprecating items, so they can
be scheduled for removal at a later stage.

A deprecation consists of adding a deprecation note
to help texts, and returning a SQL warning upon
the use.

Most of the affected regression tests have been
changed to use new or alternative syntax.  A few tests,
in particular in the 'sys_vars' suite are still run
with the deprecated syntax, to verify that the options 
still work, and give the correct warning text.

Some deprecation warnings introduced by an earlier patch
included the MySQL version number when the feature/option
would be removed.  Since future version numbers may not
be known, the text has been changed to "and will be removed 
in a future release".

The following items have been adressed by this patch:
-----------------------------------------------------
--master-retry-count       option deprecated (see also Bug#44209)
--languages | -l           option deprecated (see also WL#751)
@@storage_engine           sysvar deprecated (see also WL#4738)
@@sql_big_tables           sysvar deprecated
@@sql_low_priority_updates sysvar deprecated
@@thread_concurrency       sysvar deprecated (see also Bug#55001)

In addition, the following is mentioned in the worklog,
but already adressed by other patches:
--------------------------------------------------------
@@table_lock_wait_timeout  sysvar was removed as part of Bug#45225

@@optimizer_search_depth ( value 63) was deprecated in WL#5252 
                              and removed as part of WL#5369