Bug#54569 Some options are not allowed to take argument when passed with loose- prefix
Boolean options cause parsing failures when they are given
with prefix loose- and an argument, either in the command
line or in configuration file.

The reason was a faulty logic which forced the parsing
to throw an error when an argument of type NO_ARG was
used together with an argument which has been identified
as a key-value pair. Despite the attribute NO_ARG these
options actually take arguments if they are of type
BOOL.


include/my_getopt.h:
  * More comments to help future refactoring
mysys/my_getopt.c:
  * removed if-statement which prevented logic for handling boolean types with arguments to be executed.
  * Added comments to aid in future refactoring.