Bug #57589 SHOW SLAVE STATUS doesnt show err:1665 on NM-OS when slave can't handle checksum
OS can't connect to the checksumming-ON NM and an error goes out to the
error log (which is good), but OS can't stop trying to reconnect
constantly (which is not). 

While practically this scenario must be pretty rare it's possible
to fix the issue a rather nice way.
Master sends back to checksum-unaware OS the ER_MASTER_FATAL_ERROR_READING_BINLOG 
critical error accompanied with a verbose clarification mentioning
the checksum situation like in the following snippet of the error log from the patch testing:

101026 20:15:45 [ERROR] Error reading packet from server: Slave can not handle replication events with the checksum that master is configured to log ( server_errno=1236)
101026 20:15:45 [ERROR] Slave I/O: Got fatal error 1236 from master when reading data from binary log: 'Slave can not handle replication events with the checksum that master is configured to log', Error_code: 1236

Master also logs a warning

101026 20:15:33 [Warning] Configured to log replication events with checksum Master rejects sending them to Slave that can not handle it.


Additional fixes for wl#2540 targeting sysvar suite.
Each system var must have a test file in there. The tests are adeed.