Bug #55790  	Read_Master_Log_Pos is wrong after Slave_IO_Error= ER_NET_PACKET_TOO_LARGE
When logging in row format, DML is written as a Table_map_event
followed by its row events. The Table_map_event is written to
the relay log by slave I/O thread even if its rows events are
discarded. But it is not displayed by MYSQLBINLOG. However,
the Read_Master_Log_Pos displayed by SHOW SLAVE STATUS is
at the end of the Table_map_Log_event.

After the patch, the header of the Table_map_log_event will
be displayed by MYSQLBINLOG so long as it was written to
the relay log even if its rows events are discarded. To
multi rows events in one statement, the headers of other
rows events also will be displayed if the last rows event
is discarded by slave I/O thread. And MYSQLBINLOG will print
a warning for the missed rows event.
Refactoring:
Mysqlbinlog prints 'at' to head cache. Log_event prints head
to the head cache and body to the body cache. Mysqlbinlog
flushes the head and body caches if needed.