The X lock is only required because purge copies the contents of the oldest view and we didn't want that view removed and destroyed while purge is busy copying. Before the patch the X lock in read_view_remove() was showing up as a high contention point via trx_commit().
The solution is to clone the oldest view under the protection of
trx_sys_t::lock in S mode and the trx_sys_t::read_view_mutex. With the S lock
we guarantee that no new transactions can be added to the trx_sys_t::trx_list
and with the mutex we guarantee that no view can be deleted.

rb://510, Approved by Jimmy Yang.